@@ -21,12 +21,21 @@ export 'dtos/keycloak_conf_dto.dart';
 export 'dtos/keycloak_token_dto.dart';
 
 /// DTO
+/// SESSION WITH TOKEN+USER INFOS
+export 'dtos/observer_session_dto.dart';
+
+/// DTO
 /// TOKENS
 export 'dtos/observer_token_response.dart';
 
 /// DTO
 /// USER INFOS
 export 'dtos/user_infos_dto.dart';
+
+/// ENUMS
+/// BUTTON TYPES
+export 'enums/observer_button_enum.dart';
+export 'enums/token_validity_status_enum.dart';
 
 /// EXCEPTIONS
 /// EXCEPTION
@@ -43,6 +52,5 @@ export 'observer_auth_feature.dart';
 /// KEYCLOAK
 export 'repositories/keycloak_repository.dart';
 
-/// SERVICES
 /// AUTH
 export 'services/observer_auth_service.dart';
