@@ -6,22 +6,20 @@ part of 'keycloak_token_dto.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$KeycloakTokenDTOImpl _$$KeycloakTokenDTOImplFromJson(
-        Map<String, dynamic> json) =>
-    _$KeycloakTokenDTOImpl(
+_KeycloakTokenDTO _$KeycloakTokenDTOFromJson(Map<String, dynamic> json) =>
+    _KeycloakTokenDTO(
       refreshToken: json['refresh_token'] as String?,
       accessToken: json['access_token'] as String?,
       idToken: json['id_token'] as String?,
-      expiresIn: json['expires_in'] as int?,
-      refreshExpiresIn: json['refresh_expires_in'] as int?,
+      expiresIn: (json['expires_in'] as num?)?.toInt(),
+      refreshExpiresIn: (json['refresh_expires_in'] as num?)?.toInt(),
       tokenType: json['token_type'] as String?,
-      notBeforePolicy: json['not-before-policy'] as int?,
+      notBeforePolicy: (json['not-before-policy'] as num?)?.toInt(),
       sessionState: json['session_state'] as String?,
       scope: json['scope'] as String?,
     );
 
-Map<String, dynamic> _$$KeycloakTokenDTOImplToJson(
-        _$KeycloakTokenDTOImpl instance) =>
+Map<String, dynamic> _$KeycloakTokenDTOToJson(_KeycloakTokenDTO instance) =>
     <String, dynamic>{
       'refresh_token': instance.refreshToken,
       'access_token': instance.accessToken,
