@@ -6,10 +6,9 @@ part of 'user_infos_dto.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$UserInfosDTOImpl _$$UserInfosDTOImplFromJson(Map<String, dynamic> json) =>
-    _$UserInfosDTOImpl(
+_UserInfosDTO _$UserInfosDTOFromJson(Map<String, dynamic> json) =>
+    _UserInfosDTO(
       sub: json['sub'] as String,
-      siren: json['siren'] as String,
       emailVerified: json['email_verified'] as bool,
       name: json['name'] as String,
       preferredUsername: json['preferred_username'] as String,
@@ -17,12 +16,12 @@ _$UserInfosDTOImpl _$$UserInfosDTOImplFromJson(Map<String, dynamic> json) =>
       givenName: json['given_name'] as String,
       familyName: json['family_name'] as String,
       email: json['email'] as String,
+      siren: json['siren'] as String?,
     );
 
-Map<String, dynamic> _$$UserInfosDTOImplToJson(_$UserInfosDTOImpl instance) =>
+Map<String, dynamic> _$UserInfosDTOToJson(_UserInfosDTO instance) =>
     <String, dynamic>{
       'sub': instance.sub,
-      'siren': instance.siren,
       'email_verified': instance.emailVerified,
       'name': instance.name,
       'preferred_username': instance.preferredUsername,
@@ -30,4 +29,5 @@ Map<String, dynamic> _$$UserInfosDTOImplToJson(_$UserInfosDTOImpl instance) =>
       'given_name': instance.givenName,
       'family_name': instance.familyName,
       'email': instance.email,
+      'siren': instance.siren,
     };
